requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
requests>=2.31.0
pyyaml>=6.0
pillow>=10.0.0
openai>=1.0.0
lxml>=4.9.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
